+++ rickled/__init__.py
@@ -272,7 +272,7 @@ class BaseRickle:
         self._input_type = None
         self._allowed_chars_pat = re.compile('[^a-zA-Z_]')
         self._keys_map = dict()
-        self._path_sep = os.getenv("RICKLE_PATH_SEP", init_args.get('RICKLE_PATH_SEP', "/"))
+        self._path_sep = init_args.get('RICKLE_PATH_SEP', os.getenv("RICKLE_PATH_SEP", "/"))
 
         self._init_args = init_args